pyproject.toml
[project]
name = "mf-backtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.3",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "scipy>=1.13.1",
    "yfinance>=1.2.0",
]
